applied c++ coding rules fixed build warnings

+++ msf_tizen_client/test/main.cpp
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2016 Samsung Electronics Co., Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #include "Search.h"
 #include "Service.h"
 #include "Clients.h"
@@ -339,7 +355,6 @@ void display_service_list()
 
 void rrrr()
 {
-
 	/*
 	char c;
 	cout << "\nEnter number to connect : \n";
@@ -542,7 +557,7 @@ void connect()
 	}
 
 	std::list<Service>::iterator service = services.begin();
-	for (int i = 0; i < n-1; i++) {
+	for (int i = 0; i < (int)n-1; i++) {
 		service++;
 	}
